Call flushLock.finishedRevision from onRevisionToSKip.
My justification here is a little hazy... but if we're being told to
skip certain revisions, we can't cause any harm to bump the FlushLock up
to that point.

And crucially, things don't work without this. After we do format
detection and load the existing database state, this is the only way
that the SupervisingDriver tells the FormatDriver what the current
version number is.

This is pretty weird. The FormatDriver shouldn't need to be told this:
it just figured it out on its own in loadAllState!