Fix: onConnectionSucceeded: publish before submit.
Submitting a downstream change can trigger hooks which call driver
methods. During an outage, it we do this on the ChangeReceiver thread,
which is the very thread that repairs the FormatDriver after an outage,
then we end up with a kind of deadlock where this thread is waiting for itself.
The result is user-visible errors for operations that ought to succeed.

We initially delayed publishing a new FormatDriver until after
submitting downstream because it felt safer to ensure everything worked
before we published the shiny new driver. However, on further
reflection, it seems nothing can actually go wrong if we publish it
beforehand. More explanation in the code comments.