Fix ReplicaSet.initialRoot.
Identify (fairly arbitrarily) one of the replicas as the "primary" and
use its state as the initialRoot for other bosks. The primary will be
the bosk whose driver is constructed first using BroadcastDriver.
This will usually correspond to the bosk whose state gets updated first,
making it a good choice to supply the initial root.