WIP Refactor: remove type parameter from BoskDriver.
This is just the automated refactor step. This commit doesn't build.

It was used only to constrain the return value from initialRoot, but
with further development, it has become clear that the costs outweigh
the benefits. It's hard to imagine an actual bug that this would move
from runtime to compile time, and even if there is one, this happens
during initialization so it would be caught very early.