Encapsulate simpleDriver.
Exposing simpleDriver in a way that encourages users to use method
reference syntax (`Bosk::simpleDriver`) limits our ability to change the
implementation.

BoskDriver and driver stacks in general is an area we definitely expect
to change over time, so it's important to encapsulate this syntactically
as a method call so that users don't need to change their code as driver
stacks evolve.