Merge bosk-bson back into bosk-mongo.
After trying to use bosk-bson outside bosk-mongo, it is very much not as
useful as I had thought. The BSON representation is really useful only
for MongoDB's specific semantics, where field order is preserved; and,
moreover, parsing JSON into BSON is an extraneous step when Jackson
already offers a mutable JsonNode representation of JSON.

I'll leave it in its own package for now, to make it easier to split out
again if we ever have a need to do so.

Merging this back before publishing a release us from ending up with a
permanent bosk-bson library out in the wild for now.