Initial TaggedUnion.
I was hoping not to need this, and let the tagged union just be
represented by the common supertype of all the variants. However, this
leads to a situation where a given object (say, Variant1) must be
serialized differently as a tagged union versus as one of the possible
values. The tagged union might be { "variant1": { "field": "value" } },
while the Variant1 object itself would just be { "field": "value" }.

This turned out to be too confusing for serialization. It's particularly
true that Jackson likes to be able to serialize an object in isolation,
given nothing but the object, and BSON had a similar trouble (which is
where I actually discovered this problem). It's possible to work around
all these problems, but in the end, it's simpler if a given object has
just one correct serialized form.

Therefore, we now need a wrapper object to represent the tagged-union
serialized form. TaggedUnion plays this role.