Automatic boxing/unboxing.
Bosk needs this so that References to primitives work properly:
a reference to an int field would be a Reference<Integer>,
and if we don't do automatic boxing, that will get a needless error when
it asks for an Integer.

It seems reasonable that this ought to be built in.